requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
Django==4.2.18
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.7.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
ipwhois==1.3.0
jwcrypto==1.5.5
myinfo==1.1.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.1.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
responses==0.23.3
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
urllib3==2.3.0
whois==1.20240129.2
